requirements.txt
autopep8==1.5.2
bookkeeper @ git+https://github.com/SIIS-Sentinel/bookkeeper
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
flake8==3.8.2
idna==2.9
importlib-metadata==1.6.0
-e git+https://github.com/SIIS-Sentinel/intruder.git#egg=intruder
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
netifaces==0.10.9
paho-mqtt==1.5.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-nmap==0.6.1
requests==2.24.0
SQLAlchemy==1.3.17
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
zipp==3.1.0
